requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
sqlalchemy
psycopg2
alembic
python-jose
passlib[bcrypt]
python-multipart
gunicorn
pydantic-settings